added function to remove token from list
