+++ Makefile
@@ -47,7 +47,8 @@ SRC 		:= 	main/minishell.c 			\
 				parser/parser.c 			\
 				parser/parser_utils.c		\
 				parser/tokenizer.c 			\
-				parser/tokenizer_utils.c	\
+				parser/tokenizer_utils1.c	\
+				parser/tokenizer_utils2.c	\
 				parser/syntax_checker.c 	\
 				parser/expand_variable.c	\
 				parser/redirection_checker.c\
